Middlewares, Politiques, Directives Blades et Achèvement de la Gestion des Transferts

+++ app/Http/Controllers/Manager/AllTransfertsController.php
@@ -6,7 +6,7 @@ use App\Models\Document;
 use App\Models\DemandeTransfert;
 use Illuminate\Contracts\View\View;
 use App\Http\Controllers\Controller;
-use App\Policies\Manager\AllTransfertPolicy;
+use App\Models\BordereauTransfert;
 use Illuminate\Http\RedirectResponse;
 
 class AllTransfertsController extends Controller
@@ -17,14 +17,19 @@ class AllTransfertsController extends Controller
 
     public function all(): View
     {
-        /* $this->authorize('all'); */
+        $this->authorize('all', DemandeTransfert::class);
         return view('manager.all-demande-transfert.transferts');
     }
 
     public function one(string $slug, DemandeTransfert $transfert): View | RedirectResponse
     {
-        /* $this->authorize('one', $transfert); */
-        if($transfert->transfere){
+        $this->authorize('one', $transfert);
+        if(empty($transfert->documents->where('archive', 0)->toArray())){
+            return redirect()
+                ->route('manager.transfert.all')
+                ->with('error', 'La Demande de Transfert ne contient aucun document non archivé');
+        }
+        if($transfert->transfere && !empty($transfert->documents->toArray())){
             if($slug !== $transfert->getSlug()){
                 return to_route('manager.transfert.show', [
                     'slug' => $transfert->getSlug(),
@@ -35,43 +40,98 @@ class AllTransfertsController extends Controller
                 'transfert' => $transfert
             ]);
         }
-        return redirect()->route('manager.all-transferts.index');
+        return redirect()->route('manager.transfert.all');
     }
 
     public function removeForCentralTranfer(Document $document, DemandeTransfert $transfert): RedirectResponse
     {
-        /* $this->authorize('removeForCentralTranfer', $transfert); */
+        $this->authorize('removeForCentralTranfer', $transfert);
         $document->update([
             'demande_transfert_id' => null
         ]);
         return redirect()
-            ->route('manager.all-transferts.show', ['slug' => $transfert->getSlug(), 'transfert' => $transfert])
+            ->route('manager.transfert.one', ['slug' => $transfert->getSlug(), 'transfert' => $transfert])
             ->with('success', 'Le Document a bien été retiré de la Demande de Transfert');
     }
 
-    public function death(DemandeTransfert $transfert): RedirectResponse
+    public function off(DemandeTransfert $transfert): RedirectResponse
     {
-        /* $this->authorize('death', $transfert); */
-        self::alert($transfert);
-        if($transfert->transfere && empty($transfert->documents->toArray())){
-            $transfert->delete();
+        $this->authorize('off', $transfert);
+        if($transfert->valide){
             return redirect()
-                ->route('manager.all-transferts.index')
-                ->with('success', 'La Demande de Transfert  a bien été supprimé');
+                ->route('manager.transfert.all')
+                ->with('error', 'La Demande de Transfert  a déjà été validée');
         }
-        return redirect()->route('manager.all-transferts.index');
+        if($transfert->transfere && !$transfert->valide){
+            $transfert->delete();
+            foreach($transfert->documents->pluck('id') as $documentId){
+                $document = Document::find($documentId);
+                $document->update([
+                    'archive' => 0,
+                    'demande_transfert_id' => null
+                ]);
+            }
+            return redirect()
+                ->route('manager.transfert.all')
+                ->with('success', 'La Demande de Transfert  a bien été annulé et les documents ont été restitué');
+        }
+        return redirect()->route('manager.transfert.all');
+    }
+
+    public function showBordereauForm(DemandeTransfert $transfert)
+    {
+        $this->authorize('showBordereauForm', $transfert);
+        if($transfert->documents->count() > 0){
+            return view('manager.all-demande-transfert.bordereau-form', [
+                'transfert' => $transfert
+            ]);
+        }
+        return redirect()
+            ->route('manager.transfert.all')
+            ->with('error', 'La Demande de Transfert ne contient aucun document');
     }
 
     public function accept(DemandeTransfert $transfert)
     {
-        /* $this->authorize('accept', $transfert); */
+        $this->authorize('accept', $transfert);
+        if($transfert->documents->count() > 0){
+            request()->validate(['observation' => ['required']]);
+            $transfert->update([
+                'valide' => 1
+            ]);
+            BordereauTransfert::create([
+                'observation' => request()->observation,
+                'demande_transfert_id' => $transfert->id
+            ]);
+            return view('manager.all-demande-transfert.bordereau-preview', [
+                'transfert' => $transfert
+            ]);
+        }
+        return redirect()
+            ->route('manager.transfert.all')
+            ->with('error', 'La Demande de Transfert ne contient aucun document');
+    }
+
+    /* INUTILES */
+
+    public function death(DemandeTransfert $transfert): RedirectResponse
+    {
+        $this->authorize('death', $transfert);
+        self::alert($transfert);
+        if($transfert->transfere && empty($transfert->documents->toArray())){
+            $transfert->delete();
+            return redirect()
+                ->route('manager.transfert.all')
+                ->with('success', 'La Demande de Transfert  a bien été supprimé');
+        }
+        return redirect()->route('manager.all-transferts.index');
     }
 
     public static function alert(DemandeTransfert $transfert)
     {
         if(!empty($transfert->documents->toArray())){
             return redirect()
-                ->route('manager.transfert.index')
+                ->route('manager.transfert.all')
                 ->with('error', 'Finalisez les Opérations pour cette Demande de Transfert avant de procéder à sa suppression');
         }
     }
